Changed all conversions back to floating point math. Fixed issue with bouncing by setting the check and response scaling of the images to an offset of 1. Also, the incrSpeed in the gamedata determines how fast the player will move against the wall (may want to take a look at changing how this is implemented).

+++ player.h
@@ -18,8 +18,8 @@ public:
   const SubjectSprite* getPlayer() const { return &player; }
 
   const std::string& getName() const { return player.getName(); }
-  int getX() const { return player.getX(); }
-  int getY() const { return player.getY(); }
+  float getX() const { return player.getX(); }
+  float getY() const { return player.getY(); }
   const Image* getImage() const { 
     return player.getImage();
   }
